Move getLevel and getMoney routes, add getBotInfo and getUserData routes
Deleted getLevel.js and getMoney.js from src/api/routes and combined them into getUserData.js. Introduced getBotInfo.js to handle requests about bot-related info such as bot stats, server stats, bot features, and bot version. Updated src/index.js export to include the debug function as it's used in getBotInfo.js to fetch bot data.
