+++ src/index.js
@@ -226,3 +226,5 @@ process.on('exit', exitHandler.bind(null));
 process.on('SIGINT', exitHandler.bind(null));
 process.on('SIGUSR1', exitHandler.bind(null));
 process.on('SIGUSR2', exitHandler.bind(null));
+
+module.exports = {client, debug}
